Initial support of zulip api [wip]

+++ core/network/src/main/java/ru/snowadv/network/model/UserPresenceDto.kt
@@ -8,7 +8,7 @@ data class AllUsersPresenceDto(
     @SerialName("presences")
     val userEmailToPresenceSources: Map<String,PresenceSourcesDto>, // user email to presence sources
     @SerialName("server_timestamp")
-    val serverTimestamp: Long,
+    val serverTimestamp: Double,
 )
 
 @Serializable
@@ -45,5 +45,5 @@ data class UserPresenceDto(
     If the timestamp is more than a few minutes in the past, the user is offline.
      */
     @SerialName("timestamp")
-    val timestamp: Long,
+    val timestamp: Double,
 )
